Create parent directory and scaffold project files
Creates the `lofmand-advertisements` directory to hold all of the
project files for the final project.

Scaffolds the index.html file, and creates empty javascript files
for each visualization, and for the style-sheet.

index.html contains empty blocks with BEM style classes and
utility IDs for placing the visualizations.  Currently uses figure
tags to contain each visualization.