Implement label on hover
Removes old lable code and seperates markup for labels into a
seperate legend group.  Significantly simplifies the code to
handle updating the legend as the mouse hovers over each slice.

On hover, displays the percentage over the centroid of each
slice, and displays the category name, the count, and the topic
words in the legend off to the right.